pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "lc29h"
dynamic = ["version"]
description = "use of the LC29H GPS module on a Raspberry Pi"
authors = [
    { name = "Sébastien Allgeyer", email = "sebastien.allgeyer@gmail.com" }
]
dependencies = [
    "pyserial",
    "pynmeagps",
    "pyrtcm"
]

[project.urls]
homepage = "https://github.com/seballgeyer/LC29H_python"

[project.scripts]
your_command = "lc29h/script/get_stream:main"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "dirty-tag"
version_file = "lc29h/__version__.py"